feat: add task for scheduled sync of repos
